pyproject.toml
[project]
name = "hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk>=5.3.0",
    "fastapi>=0.121.3",
    "ffmpeg-python>=0.2.0",
    "google-genai>=1.52.0",
    "httpx>=0.28.1",
    "langgraph>=1.0.3",
    "librosa>=0.11.0",
    "numpy>=2.3.5",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "soundfile>=0.13.1",
    "streamlit>=1.51.0",
    "twilio>=9.8.7",
    "unicorn>=2.1.4",
    "websockets>=15.0.1",
]
